Move factories file to tests root
